+++ isis/src/base/objs/TrackingTable/TrackingTable.cpp
@@ -129,7 +129,7 @@ namespace Isis {
     if (index >= (unsigned int)m_fileList.size()) {
       QString msg = "Cannot convert pixel [" + toString(pixel)
                   + "] to a filename, pixel is above valid maximum ["
-                  + toString(VALID_MINUI4 + m_fileList.size()) + "].";
+                  + toString((int)(VALID_MINUI4 + m_fileList.size())) + "].";
       throw IException(IException::Programmer, msg, _FILEINFO_);
     }
 
@@ -179,7 +179,7 @@ namespace Isis {
     if (index >= (unsigned int)m_fileList.size()) {
       QString msg = "Cannot convert pixel [" + toString(pixel)
                   + "] to a serial number, pixel is above valid maximum ["
-                  + toString(VALID_MINUI4 + m_fileList.size()) + "].";
+                  + toString((int)(VALID_MINUI4 + m_fileList.size())) + "].";
       throw IException(IException::Programmer, msg, _FILEINFO_);
     }
